MVC - Observer with Timer  - TimerTask within Chronometer could be a design error because it doesnt allow task reutilization

+++ src/tomatron/model/IChronometer.java
@@ -1,6 +1,7 @@
 package tomatron.model;
 
 public interface IChronometer extends Observable{
+
 	enum STATE { 
 		STOPPED("stopped"),
 		RUNNING("running"),
@@ -24,8 +25,6 @@ public interface IChronometer extends Observable{
 	void stop();
 	void setMinutes(int minutes);
 	int getMinutes();
-	int getCurrentTime();
-
+	double getCurrentTime();
 	STATE getState();
-
 }
